refactor(openomni): rename wait service module to lifecycle
dep-check golden-principle gate rejects catch-all filenames; the module
owns the Wait lifecycle effect transitions, so it is named for that.

Roadmap: #215
