+++ packages/openomni/src/wait/lifecycle.ts
